neurons -> n_neurons and neuron_type
This has an impact on all user-facing code (tests, examples),
so this commit touches lots of files. The underlying object and builder
changes should be pretty straightforward. In general, we're moving
towards only using Ensemble, Node, Network, Connection and Probe
as "first level" objects, and Neurons, Synapses and LearningRules
as "second level" objects. The second level objects shouldn't
be used as keys into signal dictionaries and the like.
This does this work for Neurons; we'll keep doing this for learning
rules and the like next.