Add `logger` fixture for saving test logs
And a corresponding py.test `--logs` command line option.

This is a `Recorder` subclass that allows for saving logged text
to files in a specific subdirectory. Like `--plots` and `--analytics`,
the directory can be specified at the command line.

The implementation is similar to `--plots`, in that the `logger`
fixture is a `logging.Logger` instance that should be used just like
any other instance. If logs are to be saved, then the log statements
(INFO and above) will be saved to a file in the specified directory.
If logs aren't saved, then the logger will be mocked out like plt.

Other changes in this commit:
- Made builder log messages DEBUGs so they don't show up in these files
- Removed all of the `if __name__ == "__main__"` blocks in tests.
  These were less useful than just calling `py.test` on those files,
  and so we should recommend that instead.
- Changed all tests that log messages to use the `logger` fixture.
- Fixed a bug in `test_examples` that occurs when a
 `.ipynb_checkpoints` folder exists in the `examples` folder.
- Some simplifications to `nengo.log` and changes to its corresponding
  test, as these tests were causing the `logger` fixture to fail.