Move unit test information to docs
And a bunch of other consolidation.
Previously, we had a lot of information (installation related, mostly)
duplicated in many places. This puts it all in one place, and points
to that place where we used to duplicate the info.

Importantly, this gives us a single place we can point people to
for how to run unit tests, meaning that we can remove the
`if __name__ == "__main__"` blocks in tests.

Co-authored by: Sean Aubin <seanaubin@gmail.com>