Make sure Operator strings are consistent
Now, all Operators accept at `tag` argument which is part of the
string representation.

Also added a few sections to the `strings` notebook to demonstrate
the Signal and Operators string representations, as well as some
methods of comparing two sims by looking at the string representations
of all of the operators.

Also added some labels for unlabelled nodes in networks.