Update to newer pytest
This involves moving `conftest.py` to the root `nengo` directory,
and removing all of the conftests from test directories; pytest will
therefore be able to find them regardless of what subdirectory they
exist in.

The way we were doing it before fails in pytest>2.7.3. It should get
fixed eventually, but with pytest 2.9 released, I suspect it's not
an issue that many projects have, so we should probably be the ones
to change in this case.

Co-authored by: Eric Hunsberger <erichuns@gmail.com>