Simplify setup.py using pytest-runner
This essentially implements the PyTest class that we previously had.
In order to run it as `python setup.py test`, we set an alias in
`setup.cfg`. However, doing so confuses setuptools because now the
`[pytest]` section in `setup.cfg` refers to the command, so it will
be incorrectly interpreted. The fix is to move the `[pytest]` section
to `tox.ini` rather than `setup.cfg`. Since we have a `tox.ini`
regardless, this feels like no biggie to me.

This also includes a few other improvements to a few files:
- Ignore the .eggs directory created by setuptools
- MANIFEST.in now includes .yml files, suggested by check-manifest
- Author information changed to ABR
- Added trove classifiers