Remove parent tracking from vocabulary.
The only thing this was used for is to return the identity transform
from transform_to which might not be correct in all cases because one
could add new vectors to a subvocab. Within the action system this
requires users to be explicit about whether to use the identity
transform (`reinterpret` cast) or produce a transform matrix which is
not the identity but can handle new vectors (`translate` cast).