Fix ThresholdingArray utility value.
When not doing a step function decoding the utility values would be lower
by the threshold than they actually are. However, this is not easy to fix
when allowing different thresholds for the ensembles in the
ThresholdingArray. But this functionality doesn't seem to be widely used
and can be emulated by feeding in bias currents which then makes the
shift on the output explicit.