Add tests for examples and install as package data
The location of the examples is not easily discoverable by the user, so
it does not hurt much if they are hidden in a zip/egg file. At a later
point a script should be added to extract them (and documentation) to a
user specified location.

I fixed the examples to pass the test, but at a later point I have still
to go through all of them and make sure they adhere to the new nengo_spa
syntax and all the descriptions make sense.