Add sp_close function for testing.
In a lot of tests it is necessary to check if the output is close to a
specific semantic pointer within a certain time interval. This function
simplifies this. Tests can be changed to use this function as tests are
edited (i.e., I did not go through all the tests and changed them).