Make Input more similar to classic nodes.
I think this is better because it is more consistent with how the SPA
system works in other places, i.e. the user does not need to learn a
new syntax for providing inputs. An Input is now instantiated like a
normal Nengo Node by providing the output function. It can be
connected up like any other SPA module with action rules.

This makes it slightly more effort to define many Inputs based on
functions as for each function a new input needs to be created and
connected, but as this as the same amount effort that is required in
standard Nengo, it should be fine. For static inputs, having an extra
input object is unnecessary anyways (even with the previous
implementation) because those static inputs can be directly provided
in action rules.

I also like that this makes it explicit which vocabulary is used to
parse the input. Before the same input function could produce
different vector depending on what module the input was provided to
(there was actually a test for this that I removed).

Finally, this simplifies the implementation as Input becomes just a
normal nengo_spa.Network.

Two features are still missing in this commit:
* Processes providing semantic pointer inputs.
* Input functions that can take in an input from the network.