Do not raise error on duplicate SPA assignment
This error was introduced in bc24bd71f26fb1f4d60e15e0066ce17f29a01f83,
but it is not clear anymore for what reason (there is probably an issue
explaining it, but it is not linked in the commit message). It seems to
me that this error has become unnecessary because nengo_spa uses less
magic now and a double assignment should not introduce any problems
anymore. Furthermore, the moment when SPA objects is explicit due to
the build() call on the Actions objects which should make it clear that
whatever is assigned to the network's attributes at that point will be
used.