Remove special setattr handling in Network.
After removing the duplicate assignment check it was only used to
automatically assign labels to networks. We don't do this in normal
Nengo, it is somewhat magic, and it is easy to manually set the label.
By removing this functionality, we can simplify the code quite a bit,
because the custom __setstate__ implementation and _initialized flag
are no longer required.