Inspect locals in action parsing.
This allows to access all variables defined in the Python code that
are in scope. That can be helpful to specify vocabularies or in the
future solvers for vocab translation. It also makes it so that
exactly the same names are used to access things in the Python code
and in the action rules (previously the network name had to be left out
in the action rules). Moreover, this completely eliminates the need to
assign SPA modules/networks as attributes.

There might be a slight issue: We allow to provide Semantic Pointer
names, but such a name can be hidden by the something in the Python
namespace. There isn't really a good way to figure out whether a name
should refer to a Semantic Pointer or a Python object as long as we do
not introduce a special syntax for Semantic Pointers.