Fix images in notebooks in doc.
This commit removes the extraction script for the notebooks.

Revert "Add cmd interface to extract examples."
This reverts commit a74b72af44e0c5da0257a4f4e847ccdf5edab367.

Decided to not include examples, docs, or tests in the installed
packages. Docs and examples can accessed online and downloaded from
there which is easier to discover than how to extract these things
from an installed package. The tests are only required for development
which will be done on a git clone rather than an installed package.