Allow non-SPA networks as root network.
Closes #174.

Instead of providing an error when accidentally using nengo.Network
and not a spa.Network, this makes it possible to use a nengo.Network.
The main reason the spa.Network was required, was to track the default
vocabularies. This commit adds a dictionary to that allows to store
the default vocabularies associated with root networks.

This is easier than figuring out whether we shoult raise an Exception
(because spa.Network is allowed to be used in nengo.Network by design).