Fix creation of unnecessary dot network.
This unnecessary network was created for the dot product between
a fixed SemanticPointer instance and a dynamic node in the AST.

This also fixes a problem when using the fixed SemanticPointer
instance on the left-hand side in the operation returning a NumPy
array instead of an AST node.

Also see #202 where this problem was first noted.