Drop support for Python 2.7 and 3.4.
These versions are not supported by the pytest-plt package. In a
few months Python 2.7 will reach end of life anyways, so it is
not worth the extra effort to keep the support.

This commit only changes the setup.py and CI test settings. It does
not remove any compatiblity code that may still be in the code base.