Disable AppVeyor
Windows builds are not that essential to Nengo SPA as the code base is
pretty platform independent. AppVeyor is also a considerable hassle
as builds are often slow (at least twice as slow as Travis-CI) and hang
in queue for long durations. Thus, feedback is slowed down considerably
and PR throughput is impacted negatively.