Continue to allow duck-typing for algebras
However, deprecate this feature to be able to remove the additional
code for this later and make AbstractAlgebra a pure ABCMeta class.
This allows to do type checking in the vocabulary without a breaking
change.

Due to the deprecation, I consider this worth a minor release and
changed the changelog accordingly.

squash! Continue to allow duck-typing for algebras

Satisfy flake8