Generate better SemanticPointer names
Instead of immediately constructing a string for the SemanticPointer
name, this uses the expr_tree classes to construct a syntax tree.
This enables keeping track of operation precedence and allows to
render a name with only the necessary parenthesis increasing the
readability of SemanticPointer names.

Furthermore, this implements a visitor for this syntax tree to limit the
rendered length while preserving correctness of the parenthesis.

Closes #255.